requirements.txt

hikari-lightbulb==2.2.4
git+https://github.com/thesadru/hikari@dd88abdd84fb5abd398dea6292300eb035a57e46
python-dotenv==0.19.0
aiohttp
asyncpg
asyncache==0.1.1
cachetools==4.2.2
APScheduler==3.8.1
asyncpraw==7.5.0
colorama==0.4.4
pyparsing
pyyaml
python-Levenshtein
types-PyYAML
pandas
seaborn
dataenforce
pytimeparse
colormap
humanize
easydev
asyncpraw